requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
beanie==1.26.0
pydantic==2.10.2
email-validator>=2.0
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
bcrypt>=4.0,<5
python-multipart==0.0.17
pymongo>=4.9,<4.10
motor==3.6.0
